create-rootfs.sh: fix configuring interfaces when no cable is connected
The "allow-hotplug" option is more reliable as it tries
to configure an interface on receiving an event from
hotplug system, udev, etc. SO, it won't try to blindly
try to bring the interface up without a cable for instance.

This also avoids the issue on having the networking
service fails because one interface could not be
configured when a cable is disconnected.

On failing, no dhclient daemon for any interface
would run resulting in DHCP issuesm such as the DHCP
dropping IP for that host after the lease time.