crate-rootfs.sh: add systemd path condition to wait for NFS mount to complete
This is not the most elegant solution, but sometimes
NFS takes a while (e.g., few seconds) to mount after
boot. This will cause our main application boot-container-apps
to fail and none of the applications will start.

By checking a script file available at homefs we will
wait for NFS to complete.